Small workaround for non-dash attributes and default action.
If no action is specified and query non-dash attributes are used docopt
will interpret the 'resource' as an action and non-dash attributes as
the the 'resource'.

For example:

stormpath applications name=test

This will set the action to 'applications' and the resource
to 'name=test'. We catch this and move the 'name=test' to the attributes
key in the arguments dict and null out the resource key. The logic
further down can then infer again, that we wanted the default list
action and what the desired resource is, in this case 'applications'.